Replaced Gradle with SBT

+++ common/src/main/java/io/github/marad/lychee/common/messages/InitialStateMessage.java
@@ -11,7 +11,7 @@ public class InitialStateMessage implements Message {
     }
 
     @Override
-    public short getType() {
+    public int getType() {
         return MessageType.STATE.code();
     }
